+++ tsplib95/distances.py
@@ -98,6 +98,27 @@ def pseudo_euclidean(start, end, round=utils.nint):
     return distance
 
 
+def xray(start, end, sx=1, sy=1, sz=1, round=utils.icost):
+    """Return x-ray crystallography distance.
+
+    :param tuple start: 3-dimensional coordinate
+    :param tuple end: 3-dimensional coordinate
+    :param float sx: x motor speed
+    :param float sy: y motor speed
+    :param float sz: z motor speed
+    :return: distance
+    """
+    if len(start) != len(end) or len(start) != 3:
+        raise ValueError('start and end but be 3-dimensional')
+
+    dx = min(abs(start[0] - end[0]), abs(abs(start[0] - end[0])) - 360)
+    dy = abs(start[1] - end[1])
+    dz = abs(start[2] - end[2])
+    distance = max(dx / sx, dy / sy, dz / sz)
+
+    return round(distance)
+
+
 TYPES = {
     'EUC_2D': euclidean,
     'EUC_3D': euclidean,
@@ -108,4 +129,6 @@ TYPES = {
     'CEIL_2D': functools.partial(euclidean, round=math.ceil),
     'GEO': euclidean,
     'ATT': euclidean,
+    'XRAY1': xray,
+    'XRAY2': functools.partial(xray, sx=1.25, sy=1.5, sz=1.15),
 }